+++ modules/hooks/combat-hooks.js
@@ -1,6 +1,5 @@
 import { SYSTEM_ID } from '../tables/constants.js';
 import { HMSocket, SOCKET_TYPES } from '../sys/sockets.js';
-import { getLastMovedDistance } from '../integrations/elevationruler.js';
 
 export class HMCombatHooks {
     static async updateCombat(combat, _roundData, _, userId) {
@@ -9,11 +8,6 @@ export class HMCombatHooks {
         const combatants = combat.turns;
         combatants.forEach((combatant) => {
             const { token } = combatant;
-
-            // elevation ruler support
-            if (foundry.utils.hasProperty(token, 'flags.elevationruler')) {
-                token.prevLastMovedDistance = getLastMovedDistance(token, combat.round);
-            }
 
             // Toggle status effects on/off based on their timers.
             const effects = combatant.actor.effects.filter((y) => y.isTemporary === true);